requirements.txt
asgiref==3.5.2
Django==4.1.2
django-phonenumber-field==7.0.0
mysqlclient==2.1.1
phonenumbers==8.12.57
sqlparse==0.4.3
tzdata==2022.5